init bootstrap

+++ ToDoApp/ToDoApp/Models/ToDo.cs
@@ -2,7 +2,7 @@
 {
     public class ToDo
     {
-        private int id;
+        private long id;
         private string title;
         private bool isUrgent;
         private bool isDone;
@@ -14,7 +14,7 @@
             this.IsDone = false;
         }
 
-        public int Id { get => id; set => id = value; }
+        public long Id { get => id; set => id = value; }
         public string Title { get => title; set => title = value; }
         public bool IsUrgent { get => isUrgent; set => isUrgent = value; }
         public bool IsDone { get => isDone; set => isDone = value; }
