+++ ToDoApp/ToDoApp/Repositories/ToDoRepo.cs
@@ -19,7 +19,7 @@ namespace ToDoApp.Repositories
             context.SaveChanges();
         }
 
-        public void Delete(int id)
+        public void Delete(long id)
         {
             ToDo thing = context.ToDos.FirstOrDefault(c => c.Id == id);
             context.Remove(thing);
@@ -31,7 +31,7 @@ namespace ToDoApp.Repositories
             return context.ToDos.ToList();
         }
 
-        public void Update(int id)
+        public void Update(long id)
         {
             ToDo thingToUpdate = context.ToDos.FirstOrDefault(c => c.Id == 1);
             context.Update(thingToUpdate);
